Import statements changed in route files : - Renamed all the import files in routes file due to changes in the controller folder structure
